+++ src/unix/jbuild
@@ -161,6 +161,7 @@
     unix_mcast_set_loop
     unix_mcast_set_ttl
     unix_mcast_modify_membership
+    unix_wait4
     not_available
   ))
   (install_c_headers (
