+++ src/unix/lwt_unix_windows.h
@@ -411,7 +411,6 @@ LWT_NOT_AVAILABLE1(unix_getpwnam_job)
 LWT_NOT_AVAILABLE1(unix_getgrnam_job)
 LWT_NOT_AVAILABLE1(unix_getpwuid_job)
 LWT_NOT_AVAILABLE1(unix_getgrgid_job)
-LWT_NOT_AVAILABLE4(unix_wait4)
 LWT_NOT_AVAILABLE1(unix_valid_dir)
 LWT_NOT_AVAILABLE1(unix_invalidate_dir)
 LWT_NOT_AVAILABLE2(unix_bind_job)
